refactor: Split up original Modmail stuff into multiple files.
